backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
asyncpg
sqlalchemy
python-dotenv
databases
pydantic-settings
psycopg2
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
aiosmtplib
langchain
langchain-chroma
langchain-community
langchain-core
langchain-ollama
langchain-text-splitters
langsmith
langchain_openai